separate urls and menu to server parts (#2211)
* extracted from index.py

* refactored index.py, moved server dependent parts to theire modules

* updated
